🚚 Move news component to page components
